requirements.txt
MarkupSafe~=3.0.3
Flask~=3.1.2
Flask-Bcrypt~=1.0.1
Flask-JWT-Extended~=4.7.1
python-dotenv~=1.1.1